ms-cafeteria/requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0

# Database
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

